Переписал логику удаления задачи

+++ src/controllers/deleteTask.php
@@ -7,20 +7,29 @@ require($_SERVER['DOCUMENT_ROOT'] . "/source/db.php");
 
 $input = json_decode(file_get_contents('php://input'));
 
-$id = (int)htmlspecialchars($input->id);
+$index = (int)htmlspecialchars($input->index);
 
 $out = array(
     'status' => 400,
     'message' => "Ошибка удаления. Непредвиденная ошибка!"
 );
 
-if(!empty($id)) {
-    $response = Task::delete($id);
+if($index >= 0) {
 
-    if($response) {
+    $id = $_SESSION['tasks_list'][$index];
+
+    if(!empty($id)) {
+        $response = Task::delete($id);
+
+        if($response) {
+            $out['status'] = 200;
+            $out['message'] = "Задача удалена!";
+        }
+    } else {
         $out['status'] = 200;
-        $out['message'] = "Задача удалена!";
+        $out['message'] = "Задача не найдена!";
     }
+    
 }
 
 echo json_encode($out);
